requirements.txt
fastapi==0.111.0
invoke==2.2.0
passlib==1.7.4
pydantic==2.7.1
pydantic_settings==2.2.1
PyJWT==2.8.0
SQLAlchemy==2.0.30
uvicorn==0.29.0
alembic==1.13.1
aiosqlite==0.20.0
pre-commit